18/08/2025

+++ src/main/java/com/ocupacional/soc/Dto/Cadastros/SetorRequestDTO.java
@@ -23,6 +23,6 @@ public class SetorRequestDTO {
     private StatusEmpresa status;
     private TipoEmpresa tipoEmpresa;
     private TipoDocumento tipoDocumento;
-    private Integer numeroDocumento;
+    private Long numeroDocumento;
 
 }
